+++ test-fixture/ios-swift-cocoapods/iOSTestApp/scenarios/LoadConfigFromFileScenario.swift
@@ -1,0 +1,26 @@
+
+import Foundation
+import UIKit
+import Bugsnag
+
+class LaunchError : Error {
+
+    init() {
+    }
+}
+
+@objc class LoadConfigFromFileScenario: Scenario {
+
+    override func startBugsnag() {
+        if let fileConfig = BugsnagConfiguration.loadConfig() {
+            fileConfig.setEndpoints(notify: config.notifyURL!.absoluteString,
+                                    sessions: config.sessionURL!.absoluteString)
+            config = fileConfig
+        }
+        Bugsnag.start(with: config)
+    }
+
+    override func run() {
+        Bugsnag.notifyError(LaunchError())
+    }
+}